Move the auth logic to a middlewares folder
We'd likely have to add more middleware + we don't like flooding the top level
