+++ src/app.ts
@@ -3,7 +3,7 @@ import connectDB from "./db.js";
 import { meetingRoutes } from "./services/meetings/meeting.routes.js";
 import { taskRoutes } from "./services/tasks/task.routes.js";
 import { dashboardRoutes } from "./services/dashboard/dashboard.routes.js";
-import { authMiddleware } from "./auth.middleware.js";
+import { authMiddleware } from "./middlewares/auth.js";
 
 const app = express();
