+++ src/services/dashboard/dashboard.routes.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import mongoose from "mongoose";
-import { AuthenticatedRequest } from "../../auth.middleware.js";
+import { AuthenticatedRequest } from "../../middlewares/auth.js";
 import { Meeting } from "../meetings/meeting.model.js";
 import { Task } from "../tasks/task.model.js";
 import { Types } from "mongoose";
